+++ src/types/Augmented/SkyBlock/Auction.ts
@@ -1,5 +1,12 @@
 import { Components } from "../../api";
 
-export type SkyBlockAuction = Components.Schemas.SkyBlockAuction & {
+export type SkyBlockAuction = Omit<
+  Components.Schemas.SkyBlockAuction,
+  "item_bytes"
+> & {
   bin?: boolean;
+  categories?: string[];
+  item_uuid?: string;
+  last_updated?: number;
+  item_bytes?: string;
 };